tests/wpt/tests/websockets/interfaces/WebSocket/events/017.html

<!doctype html>
<title>WebSockets: this, e.target, e.currentTarget, e.eventPhase</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<meta name="variant" content="?wpt_flags=h2">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo_raw');
  ws.addEventListener('open', function(e) {
    var this_val = this;
    t.step(function() {
      // first a text frame, then a frame with reserved opcode 3
      // which should fail the connection

      // send '\\x81\\x04test\\x83\\x03LOL' in bytes
      ws.send(new Uint8Array([129, 4, 116, 101, 115, 116, 131, 3, 76, 79, 76]));
      assert_equals(this_val, ws);
      assert_equals(e.target, ws);
      assert_equals(e.currentTarget, ws);
      assert_equals(e.eventPhase, 2);
    });
  }, false);
  ws.addEventListener('message', function(e) {
    var this_val = this;
    t.step(function() {
      assert_equals(this_val, ws);
      assert_equals(e.target, ws);
      assert_equals(e.currentTarget, ws);
      assert_equals(e.eventPhase, 2);
    });
  }, false);
  ws.addEventListener('error', function(e) {
    var this_val = this;
    t.step(function() {
      assert_equals(this_val, ws);
      assert_equals(e.target, ws);
      assert_equals(e.currentTarget, ws);
      assert_equals(e.eventPhase, 2);
    });
  }, false);
  ws.addEventListener('close', function(e) {
    var this_val = this;
    t.step(function() {
      assert_equals(this_val, ws);
      assert_equals(e.target, ws);
      assert_equals(e.currentTarget, ws);
      assert_equals(e.eventPhase, 2);
      t.done();
    });
  }, false);
});
</script>
